Change vec objects to Proxys
The idea here is to be able to restrict certain features provided by Arrays.
For example, we don't want to change the length because we already know the dimension.
If we're calling certain functions, the return value should only be a vec if it
fits the constraints of being a vec. Otherwise turn it into an array.