Drastically improve performance and add copy function for deep copies
Instead of abusing map and vecn methods, just use simple for loops.
For example, div used to call pow twice and times once to avoid calling constructors.
That was probably a bad decision in the first place, since every arithmetic operator
passed the result through a factory anyway.

times, on the other hand, only uses map. However, Array's map is already slower than
for loops, and the overloaded method here performs a check of the results in addition
to the map. So, it's always beneficial to avoid calls to map when it comes to performance.

Changes result in a 4-5x speedup on arithmetic operations.